broke _sum_counters() out into separate module
